refactor: remove `pluginsModulesPath` as a top-level option
