+++ src/runtimes/node/bundlers/esbuild/src_files.ts
@@ -1,4 +1,4 @@
-import type { GetSrcFilesFunction } from '../../../runtime'
+import type { GetSrcFilesFunction } from '..'
 import { filterExcludedPaths, getPathsOfIncludedFiles } from '../../utils/included_files'
 import { getPackageJson, PackageJson } from '../../utils/package_json'
 import { getNewCache, TraversalCache } from '../../utils/traversal_cache'
