+++ src/runtimes/node/bundlers/zisi/src_files.ts
@@ -4,9 +4,9 @@ import { dirname, basename, normalize } from 'path'
 import { not as notJunk } from 'junk'
 import precinct from 'precinct'
 
+import type { GetSrcFilesFunction } from '..'
 import { FeatureFlags } from '../../../../feature_flags'
 import { nonNullable } from '../../../../utils/non_nullable'
-import type { GetSrcFilesFunction } from '../../../runtime'
 import { filterExcludedPaths, getPathsOfIncludedFiles } from '../../utils/included_files'
 import { getPackageJson, PackageJson } from '../../utils/package_json'
 import { getNewCache, TraversalCache } from '../../utils/traversal_cache'
